requirements.txt
# =============================
# Core numerical packages
# =============================
numpy>=2,<3
scipy>=1.13,<2.0
pandas>=2.2,<3.0

# =============================
# Optuna for Hyperparameter Optimization
# =============================
optuna>=3.1,<5.0

# =============================
# Visualization & EDA
# =============================
matplotlib>=3.8,<4.0
seaborn>=0.13,<0.14

# =============================
# Machine Learning / Preprocessing
# =============================
scikit-learn>=1.4,<1.6

# =============================
# Quantum SDK + Machine Learning + TorchConnector
# (this will pull a compatible PyTorch + torchvision)
# =============================
qiskit==1.4.4
qiskit-machine-learning[torch]==0.8.4
qiskit-aer>=0.15.1

# =============================
# Alpaca API (market data/trading)
# =============================
alpaca-py>=0.15,<0.20

# =============================
# LaTeX Parsing / Encoding
# =============================
pylatexenc>=2.0,<4.0

# =============================
# YAML config parsing
# =============================
PyYAML>=6.0,<7.0

# =============================
# Optional: Notebooks and Dev Tools
# =============================
jupyterlab>=4.1,<5.0
ipykernel>=6.29,<7.0
tqdm>=4.66,<5.0

